rename rule name-casing -> name-property-casing
