+++ lib/rules/name-property-casing.js
